Material Theme (#222)
* Update FeaturesView to use Material Theme colors.

* Update features view UI with proper material design colors.

* Explicitly define styles for top app bar.

* Add step view to project.

* Implement EnumOptionList for should vault toggle.

* Implement enum option list for approve order view.

* Use EnumOptionList for SCA.

* Remove customer id input field.

* Add content area to order creation button.

* Add green state to project for action button.

* Make green more muted.

* Optionally show created order properties.

* Update Order Approval step to use new UX styule.

* Add test card button to project.

* Update styling of approve order section.

* Update StepContainer styles.

* Update ApproveOrder view UI to contain additional action state enums.

* Fix compiler errors.

* Extract approve order composables into steps.

* Update styling on column view.

* Implement Approve Order with new UX style.

* Rename StatefulActionButton to ActionButtonColumn.

* Rename StepContainer to StepHeader and add spacing constants.

* Update Card Approve Order flow.

* Start refactoring vault card view.

* Update Vault view model setup token.

* Implement vault card views with new design system.

* Update Vault flow to use compose.

* Add spacer at end of vault card view.

* Modify step header vertical alignment of step number.

* Restyle PayPalWeb view.

* Update PayPalWeb flow to use new compose layout.

* Convert PayPalNativeView to use Jetpack compose.

* Clean up CardForm lint errors.

* Clean up additional lint errors.

* Remove magic numbers where possible.

* Clean up additional lint errors.

* Remove unused strings.

* Remove StringResourceList utility.

* Remove SettingsFragment that is no longer used.

* Remove unused string resources.

* Remove approve order success view.

* Extract Context.getActivity() extention function.

* Clean up ApproveOrderView annotations.

* Cleanup unused resources.

* Remove no longer needed suppression.

* Add opt in annotation.

* Refactor PayPalSDKErrorView to be ErrorView.

* Remove unecessary files and cleanup Spacer.

* Remove unused string resources.

* Update PayPalButtons view spacing.

* Fix lint errors.

* Extract slider into it's own option view.

* Add PayPalButtons value readout to UI.

* Update app icon.

* Add additional annotation to MainActivity to resolve compilation error.

* Remove unused resources and tweak app icon.

* Tweak UI styles on Card ApproveOrder.

* Remove unused imports.

* Change padding modifier.

* Tweak SelectCardView styling.

* Make step composables private.

* Remove exclamation points from copy.

* Fix rebasing errors.

* Clean up detekt errors.

* Update PayPal Vault UI to match updated UI.

* Fix lint errors.

* Rename ActionButtonState to ActionState.

* Cleanup magic numbers.

* Refactor hoisted click listener name.

* Update FeaturesView with UIConstants.

* Update property change naming convention throughout project.

* Change ActionState.Ready to ActionState.Idle.

* Remove unecessary modifier from StepHeader class.

* Create IntSlider composable to allow for a slider with integer values.

* Cleaning up remaining magic numbers.

* Fix lint errors.

* Cleanup android lint errors.

+++ Demo/src/main/java/com/paypal/android/ui/paypalbuttons/PayPalButtonsViewModel.kt
@@ -38,7 +38,7 @@ class PayPalButtonsViewModel : ViewModel() {
             _uiState.update { it.copy(payPalButtonLabel = value) }
         }
 
-    var customCornerRadius: Float?
+    var customCornerRadius: Int?
         get() = _uiState.value.customCornerRadius
         set(value) {
             _uiState.update { it.copy(customCornerRadius = value) }
